@@ -1,188 +1,103 @@
 {
-  "name": "nursing-ux",
-  "author": "kaptest",
-  "description": "Nursing Faculty Site",
-  "main": "src/tmp-app/bootstrap.js",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/KaplanTestPrep/NursingIT.Frontend.git"
-  },
-  "scripts": {
-    "postinstall": "bower install && typings install",
-    "start": "gulp",
-    "test": "gulp test",
-    "coveralls": "node ./node_modules/coveralls/bin/coveralls.js < ./report/remap/lcov.info"
-  },
-  "keywords": [
-    "angular2",
-    "starter",
-    "seed",
-    "gulp",
-    "webpack",
-    "extensible",
-    "systemjs",
-    "systemjs builder",
-    "travis",
-    "karma",
-    "jasmine",
-    "istanbul"
-  ],
+  "name": "cli21",
+  "version": "0.0.0",
   "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/KaplanTestPrep/NursingIT.Frontend/issues"
+  "angular-cli": {},
+  "scripts": {
+    "start": "ng serve",
+    "lint": "tslint \"src/**/*.ts\"",
+    "test": "ng test",
+    "pree2e": "webdriver-manager updathe",
+    "e2e": "protractor"
   },
-  "devDependencies": {
-    "@types/bootstrap": "^3.3.32",
-    "@types/bootstrap-datepicker": "0.0.3",
-    "@types/browser-sync": "0.0.33",
-    "@types/colors": "^0.6.33",
-    "@types/connect": "^3.4.30",
-    "@types/core-js": "^0.9.34",
-    "@types/crypto-js": "^3.1.32",
-    "@types/cryptojs": "3.1.29",
-    "@types/del": "^2.2.31",
-    "@types/es6-shim": "^0.31.32",
-    "@types/glob": "^5.0.30",
-    "@types/gulp": "^3.8.32",
-    "@types/gulp-csso": "^1.0.29",
-    "@types/gulp-htmlmin": "^1.3.29",
-    "@types/gulp-if": "0.0.30",
-    "@types/gulp-inject": "0.0.30",
-    "@types/gulp-load-plugins": "0.0.28",
-    "@types/gulp-plumber": "0.0.29",
-    "@types/gulp-protractor": "^1.0.29",
-    "@types/gulp-rename": "0.0.31",
-    "@types/gulp-replace": "0.0.29",
-    "@types/gulp-rev": "^5.0.29",
-    "@types/gulp-rev-replace": "^0.2.29",
-    "@types/gulp-size": "^1.2.29",
-    "@types/gulp-sourcemaps": "0.0.29",
-    "@types/gulp-tslint": "^3.6.30",
-    "@types/gulp-typescript": "0.0.32",
-    "@types/gulp-uglify": "0.0.29",
-    "@types/gulp-useref": "^3.0.29",
-    "@types/gulp-util": "^3.0.29",
-    "@types/jasmine": "^2.2.34",
-    "@types/jquery": "^2.0.32",
-    "@types/jwt-decode": "^1.4.28",
-    "@types/karma": "^0.13.33",
-    "@types/karma-coverage": "^0.5.30",
-    "@types/karma-jasmine": "0.0.28",
-    "@types/lodash": "^4.14.36",
-    "@types/mkpath": "^0.1.29",
-    "@types/moment": "^2.13.0",
-    "@types/moment-timezone": "^0.2.31",
-    "@types/open": "0.0.29",
-    "@types/reflect-metadata": "0.0.4",
-    "@types/run-sequence": "0.0.28",
-    "@types/serve-static": "^1.7.31",
-    "@types/systemjs": "^0.19.31",
-    "@types/typeahead": "0.11.27",
-    "@types/whatwg-fetch": "0.0.31",
-    "@types/when": "^2.4.27",
-    "@types/x-editable": "^1.5.27",
-    "@types/yargs": "0.0.31",
-    "@types/zone.js": "0.0.27",
-    "bower": "^1.7.7",
-    "browser-sync": "^2.12.8",
-    "codelyzer": "0.0.19",
-    "colors": "1.1.2",
-    "connect": "^3.3.4",
-    "connect-history-api-fallback": "^1.2.0",
-    "del": "~2.2.0",
-    "glob": "^7.0.3",
-    "gulp": "^3.9.1",
-    "gulp-cssnano": "^2.1.2",
-    "gulp-csso": "1.1.0",
-    "gulp-htmlmin": "^1.3.0",
-    "gulp-if": "~2.0.1",
-    "gulp-inject": "4.1.0",
-    "gulp-jscs": "^1.6.0",
-    "gulp-jshint": "^1.11.2",
-    "gulp-load-plugins": "^1.0.0",
-    "gulp-plumber": "^1.0.0",
-    "gulp-print": "^1.1.0",
-    "gulp-protractor": "2.4.0",
-    "gulp-rename": "~1.2.0",
-    "gulp-replace": "^0.5.4",
-    "gulp-rev": "^7.0.0",
-    "gulp-rev-replace": "^0.4.3",
-    "gulp-size": "^1.2.1",
-    "gulp-sourcemaps": "^1.6.0",
-    "gulp-traceur": "0.17.*",
-    "gulp-tslint": "^5.0.0",
-    "gulp-typescript": "~2.13.4",
-    "gulp-uglify": "~1.5.3",
-    "gulp-useref": "~3.1.0",
-    "gulp-util": "^3.0.6",
-    "gulp-zip": "^3.0.2",
-    "jasmine": "2.4.1",
-    "jasmine-core": "~2.4.1",
-    "jasmine2-reporter": "0.2.0",
-    "jshint-stylish": "^2.0.1",
-    "karma": "~0.13.22",
-    "karma-chrome-launcher": "^0.1.4",
-    "karma-cli": "^0.0.4",
-    "karma-coverage": "~1.0.0",
-    "karma-ie-launcher": "^1.0.0",
-    "karma-jasmine": "~1.0.2",
-    "karma-phantomjs-launcher": "~1.0.0",
-    "karma-sourcemap-loader": "^0.3.7",
-    "live-server": "^0.9.1",
-    "mkpath": "~0.1.0",
-    "open": "0.0.5",
-    "phantomjs-prebuilt": "^2.1.7",
-    "remap-istanbul": "~0.6.4",
-    "require-dir": "~0.3.0",
-    "rework": "~1.0.1",
-    "rework-npm": "~1.0.0",
-    "robocopy": "^0.1.15",
-    "run-sequence": "^1.2.1",
-    "serve-static": "~1.8.1",
-    "systemjs-builder": "0.15.23",
-    "traceur": "~0.0.110",
-    "tslint": "^3.10.2",
-    "typings": "^1.0.4",
-    "yargs": "^4.7.1"
-  },
+  "private": true,
   "dependencies": {
-    "@angular/common": "2.0.0-rc.4",
-    "@angular/compiler": "2.0.0-rc.4",
-    "@angular/core": "2.0.0-rc.4",
-    "@angular/http": "2.0.0-rc.4",
-    "@angular/platform-browser": "2.0.0-rc.4",
-    "@angular/platform-browser-dynamic": "2.0.0-rc.4",
-    "@angular/router": "3.0.0-beta.1",
-    "@angular/router-deprecated": "2.0.0-rc.2",
-    "@angular/upgrade": "2.0.0-rc.4",
-    "angular2-in-memory-web-api": "0.0.14",
-    "angulartics2": "1.1.2",
-    "bootstrap": "^3.3.5",
+    "@angular/common": "2.2.3",
+    "@angular/compiler": "2.2.3",
+    "@angular/core": "2.2.3",
+    "@angular/forms": "2.2.3",
+    "@angular/http": "2.2.3",
+    "@angular/platform-browser": "2.2.3",
+    "@angular/platform-browser-dynamic": "2.2.3",
+    "@angular/router": "3.1.2",
+    "@types/bootstrap-datepicker": "0.0.5",
+    "@types/crypto-js": "^3.1.32",
+    "@types/jquery": "^2.0.34",
+    "@types/jwt-decode": "^1.4.28",
+    "@types/lodash": "^4.14.39",
+    "@types/moment-timezone": "^0.2.32",
+    "angulartics2": "1.1.12",
+    "bootstrap": "3.3.7",
     "bootstrap-datepicker": "1.6.1",
-    "bootstrap-sass": "^3.3.6",
-    "bootstrap-select": "1.9.3",
-    "bower": "^1.7.9",
-    "core-js": "^2.4.0",
-    "crypto-js": "3.1.5",
+    "bootstrap-select": "1.11.2",
+    "core-js": "^2.4.1",
+    "crypto-js": "3.1.7",
     "datatables": "1.10.12",
     "datatables-responsive": "1.0.7",
-    "es6-shim": "^0.35.0",
-    "font-awesome": "^4.6.1",
-    "jquery": "^2.2.3",
-    "jwt-decode": "1.1.0",
-    "lodash": "^4.13.1",
+    "jquery": "3.1.1",
+    "jwt-decode": "2.1.0",
+    "lodash": "4.16.6",
     "moment": "2.15.1",
     "moment-timezone": "0.5.6",
-    "ng2-page-scroll": "^1.2.1",
-    "reflect-metadata": "^0.1.3",
-    "rxjs": "5.0.0-beta.6",
-    "systemjs": "0.19.27",
+    "ng2-lazyload-image": "2.2.2",
+    "ng2-page-scroll": "3.1.4",
+    "rxjs": "5.0.0-beta.12",
     "tablesaw": "2.0.3",
-    "timepicker": "1.11.1",
+    "timepicker": "1.11.6",
+    "ts-helpers": "^1.1.1",
     "typeahead.js": "0.11.1",
-    "whatwg-fetch": "0.7.0",
-    "when": "3.7.3",
     "x-editable": "1.5.1",
-    "zone.js": "0.6.12"
+    "zone.js": "^0.6.23"
+  },
+  "devDependencies": {
+    "@angular/compiler-cli": "2.2.1",
+    "@types/bootstrap": "^3.3.32",
+    "@types/bootstrap-datepicker": "0.0.4",
+    "@types/browser-sync": "0.0.34",
+    "@types/crypto-js": "^3.1.32",
+    "@types/gulp": "^3.8.32",
+    "@types/gulp-load-plugins": "0.0.28",
+    "@types/gulp-rename": "0.0.31",
+    "@types/gulp-shell": "0.0.29",
+    "@types/jasmine": "2.5.38",
+    "@types/jquery": "^2.0.33",
+    "@types/jquery.datatables": "1.10.31",
+    "@types/jwt-decode": "^1.4.28",
+    "@types/lodash": "^4.14.37",
+    "@types/moment": "^2.13.0",
+    "@types/moment-timezone": "^0.2.32",
+    "@types/node": "^6.0.42",
+    "@types/open": "0.0.29",
+    "@types/run-sequence": "0.0.28",
+    "@types/typeahead": "0.11.27",
+    "@types/x-editable": "^1.5.27",
+    "@types/yargs": "^6.3.0",
+    "@types/zone.js": "0.0.27",
+    "angular-cli": "1.0.0-beta.21",
+    "bower": "1.7.9",
+    "browser-sync": "^2.12.8",
+    "codelyzer": "~1.0.0-beta.3",
+    "connect-history-api-fallback": "^1.2.0",
+    "gulp": "3.9.1",
+    "gulp-load-plugins": "^1.0.0",
+    "gulp-print": "^1.1.0",
+    "gulp-rename": "~1.2.0",
+    "gulp-shell": "0.5.2",
+    "gulp-zip": "^3.0.2",
+    "jasmine-core": "2.5.2",
+    "jasmine-spec-reporter": "2.5.0",
+    "karma": "1.2.0",
+    "karma-chrome-launcher": "^2.0.0",
+    "karma-cli": "^1.0.1",
+    "karma-jasmine": "^1.0.2",
+    "karma-remap-istanbul": "^0.2.1",
+    "open": "0.0.5",
+    "protractor": "4.0.9",
+    "require-dir": "0.3.1",
+    "run-sequence": "^1.2.1",
+    "ts-node": "1.2.1",
+    "tslint": "3.13.0",
+    "typescript": "~2.0.3",
+    "webdriver-manager": "10.2.5"
   }
 }
