@@ -1,8 +1,8 @@
-import {Injectable} from '@angular/core';
-import {Http, Headers, RequestOptions, Response} from '@angular/http';
-import {Observable} from 'rxjs/Rx';
+import { Injectable } from '@angular/core';
+import { Http, Headers, RequestOptions, Response } from '@angular/http';
+import { Observable } from 'rxjs/Rx';
 // import {AuthService} from './auth';
-import {links} from '../constants/config';
+import { links } from '../constants/config';
 // import {CommonService} from './common';
 import { AuthService } from './auth.service';
 import { CommonService } from './common.service';
@@ -13,27 +13,32 @@ export class LogService {
 
     }
 
-    error(error: string, stacktrace: string = null, reason: string =  null): any {
-        let body = JSON.stringify({
-            ErrorType: error,
-            ErrorDescription: stacktrace,
-            Reason: reason
-        });
-        let url: string = this.common.getLogServer();
-        let headers: Headers = new Headers();
-        headers.append('Authorization', this.auth.authheader);
-        headers.append('Accept', 'application/json');
-        headers.append('Content-Type', 'application/json');
-        let options: RequestOptions = new RequestOptions();
-        options.headers = headers;
+    error(error: string, stacktrace: string = null, reason: string = null): any {
+        try {
+            let body = JSON.stringify({
+                ErrorType: error,
+                ErrorDescription: stacktrace,
+                Reason: reason
+            });
+            let url: string = this.common.getLogServer();
+            let headers: Headers = new Headers();
+            headers.append('Authorization', this.auth.authheader);
+            headers.append('Accept', 'application/json');
+            headers.append('Content-Type', 'application/json');
+            let options: RequestOptions = new RequestOptions();
+            options.headers = headers;
 
-        this.http.post(url, body, options)
-            .map(res => <any>res.json())
-            .subscribe(
-                 response => console.log(response),
-                 error=>this.handleError,
-                 () => console.log('Error logged ...')
-        );
+            this.http.post(url, body, options)
+                .map(res => <any>res.json())
+                .subscribe(
+                response => console.log(response),
+                error => this.handleError,
+                () => console.log('Error logged ...')
+                );
+        } catch (error) {
+            console.log('error caught');
+        }
+
     }
 
     handleError(error: Response): any {
