@@ -1,2 +1,2 @@
-﻿var requireDir = require('require-dir');
+var requireDir = require('require-dir');
 var tasks = requireDir('./tasks');
