@@ -1,95 +1,43 @@
-module.exports = function(config) {
-  var gulpConfig = require('./gulpnew.config')();
+// Karma configuration file, see link for more information
+// https://karma-runner.github.io/0.13/config/configuration-file.html
 
-  /**
-   * List of npm packages that imported via `import` syntax
-   */
-  var dependencies = [
-    '@angular',
-    'lodash',
-    'rxjs',
-    'angulartics2'
-  ];
-
-  var configuration = {
-    basePath: './',
-
-    frameworks: ['jasmine'],
-    browsers: ['PhantomJS'],
-    reporters: ['progress', 'coverage'],
-
-    preprocessors: {},
-
-    // Generate json used for remap-istanbul
-    coverageReporter: {
-      dir: 'report/',
-      reporters: [
-        { type: 'json', subdir: 'report-json' }
-      ]
-    },
-
-    files: [
-      'node_modules/traceur/bin/traceur-runtime.js',
-      'node_modules/core-js/client/shim.min.js',
-      'node_modules/systemjs/dist/system-polyfills.js',
-      'node_modules/zone.js/dist/zone.js',
-      'node_modules/zone.js/dist/async-test.js',
-      'node_modules/zone.js/dist/fake-async-test.js',
-      'node_modules/reflect-metadata/Reflect.js',
-      'node_modules/systemjs/dist/system.src.js'
+module.exports = function (config) {
+  config.set({
+    basePath: '',
+    frameworks: ['jasmine', 'angular-cli'],
+    plugins: [
+      require('karma-jasmine'),
+      require('karma-chrome-launcher'),
+      require('karma-remap-istanbul'),
+      require('angular-cli/plugins/karma')
     ],
-
-    // proxied base paths
-    proxies: {
-      // required for component assests fetched by Angular's compiler
-      "/src/": "/base/src/",
-      "/app/": "/base/src/app/",
-      "/node_modules/": "/base/node_modules/"
+    files: [
+      { pattern: './src/test.ts', watched: false }
+    ],
+    preprocessors: {
+      './src/test.ts': ['angular-cli']
     },
-
+    mime: {
+      'text/x-typescript': ['ts','tsx']
+    },
+    remapIstanbulReporter: {
+      reports: {
+        html: 'coverage',
+        lcovonly: './coverage/coverage.lcov'
+      }
+    },
+    angularCli: {
+      config: './angular-cli.json',
+      environment: 'dev'
+    },
+    reporters: config.angularCli && config.angularCli.codeCoverage
+              ? ['progress', 'karma-remap-istanbul']
+              : ['progress'],
     port: 9876,
     colors: true,
     logLevel: config.LOG_INFO,
-    autoWatch: true
-  };
-
-  configuration.preprocessors[gulpConfig.tmpApp + '**/!(*.spec)+(.js)'] = ['coverage'];
-  configuration.preprocessors[gulpConfig.tmpApp + '**/*.js'] = ['sourcemap'];
-  configuration.preprocessors[gulpConfig.tmpTest + '**/*.js'] = ['sourcemap'];
-
-  var files = [
-    gulpConfig.tmpTest + 'test-helpers/global/**/*.js',
-    gulpConfig.assetsPath.scripts + 'systemjs.conf.js',
-    'karma-test-shim.js',
-    createFilePattern(gulpConfig.tmpApp + '**/*.js', { included: false }),
-    createFilePattern(gulpConfig.tmpTest + 'test-helpers/*.js', { included: false }),
-    createFilePattern(gulpConfig.app + '**/*.html', { included: false }),
-    createFilePattern(gulpConfig.app + '**/*.css', { included: false }),
-    createFilePattern(gulpConfig.app + '**/*.ts', { included: false, watched: false }),
-    createFilePattern(gulpConfig.tmpApp + '**/*.js.map', { included: false, watched: false })
-  ];
-
-  configuration.files = configuration.files.concat(files);
-
-  dependencies.forEach(function(key) {
-    configuration.files.push({
-        pattern: 'node_modules/' + key + '/**/*.js',
-        included: false,
-        watched: false
-    });
+    autoWatch: true,
+    browsers: ['Chrome'],
+    singleRun: false
   });
-
-  if (process.env.APPVEYOR) {
-    configuration.browsers = ['IE'];
-    configuration.singleRun = true;
-    configuration.browserNoActivityTimeout = 90000; // Note: default value (10000) is not enough
-  }
-
-  config.set(configuration);
-
-  // Helpers
-  function createFilePattern(path, config) {
-    config.pattern = path;
-    return config;
-  }
-}
+};
