@@ -1,5 +1,5 @@
 import {Component, OnInit, Input} from '@angular/core';
-import {Router, ROUTER_DIRECTIVES} from '@angular/router';
+import {Router} from '@angular/router';
 import { AuthService } from './../../services/auth.service';
 
 @Component({
@@ -11,15 +11,13 @@ import { AuthService } from './../../services/auth.service';
 			<li><a [routerLink]="['/home']" class="menu-icon home-icon" >Home</a></li>
 			<li><a [routerLink]="['/reports']" class="menu-icon reporting-icon" >View Reports</a></li>
 			<li><a [routerLink]="['/rosters']" class="menu-icon student-icon" >View Rosters</a></li>
-			<li><a  [routerLink]="['/groups']" class="menu-icon group-icon" >Manage Groups</a></li>
+			<li><a [routerLink]="['/groups']" class="menu-icon group-icon" >Manage Groups</a></li>
 			<li><a [routerLink]="['/tests']" class="menu-icon test-icon" >Manage Tests</a></li>
 			<li><a [routerLink]="['/account']" class="menu-icon manage-icon" >Manage Account</a></li>
 			<li><a href="#" class="menu-icon sign-out-icon" (click)="logout($event);">Sign Out</a></li>
 		</ul>
 	</div>
-</nav>`,
-    providers:[AuthService],
-    directives: [ROUTER_DIRECTIVES]
+</nav>`
 })
 export class DropdownMenuComponent implements OnInit {
     @Input() ariaDisabled;
