@@ -1,16 +1,14 @@
 import {Component, OnInit} from '@angular/core';
-import {Router, ROUTER_DIRECTIVES} from '@angular/router';
-import {Title} from '@angular/platform-browser';
-import * as _ from 'lodash';
+import { Router } from '@angular/router';
+import { Title } from '@angular/platform-browser';
+// import * as _ from 'lodash';
 import { AuthService } from './../../services/auth.service';
 import { PageHeaderComponent } from './../shared/page-header.component';
 import { PageFooterComponent } from './../shared/page-footer.component';
 
 @Component({
     selector: 'groups',
-    providers: [AuthService],
-    templateUrl: 'components/groups/groups.component.html',
-    directives: [ROUTER_DIRECTIVES, PageHeaderComponent, PageFooterComponent]
+    templateUrl: './groups.component.html'
 })
 
 export class GroupsComponent implements OnInit {
@@ -29,7 +27,7 @@ export class GroupsComponent implements OnInit {
 
     getLatestInstitution(): string {
         if (this.auth.institutions != null && this.auth.institutions != 'undefined') {
-            let latestInstitution = _.first(_.orderBy(JSON.parse(this.auth.institutions), 'InstitutionId', 'desc'))
+            let latestInstitution:any = _.first(_.orderBy(JSON.parse(this.auth.institutions), 'InstitutionId', 'desc'))
             if (latestInstitution)
                 return latestInstitution.InstitutionName;
         }
