@@ -2,8 +2,6 @@ import { Observable } from 'rxjs/Rx';
 import { RequestOptions, Headers, Http, Response } from '@angular/http';
 import { Injectable } from '@angular/core';
 import { AuthService } from './../../services/auth.service';
-import { ProfileModel } from './../../models/profile.model';
-
 
 @Injectable()
 export class RosterService {
@@ -50,7 +48,8 @@ export class RosterService {
     searchStudents(url): Observable<Response> {
         return this.http.get(url, this.getRequestOptionsWithEmptyBody());
     }
-    setUserPreference(url: string, input: string): Observable<Response> {
+
+     setUserPreference(url: string, input: string): Observable<Response> {
         let self = this;
         let headers: Headers = new Headers({
             'Accept': 'application/json',
@@ -66,4 +65,5 @@ export class RosterService {
     getRosterCohortUserPreference(url): Observable<Response> {
         return this.http.get(url, this.getRequestOptionsWithEmptyBody());
     }
+
 }
