@@ -1,9 +1,9 @@
 import { Component, OnInit} from '@angular/core';
-import { Router, ROUTER_DIRECTIVES } from '@angular/router';
+import { Router,  } from '@angular/router';
 import { Title} from '@angular/platform-browser';
 
 // import { AuthService } from '../../services/auth';
-import * as _ from 'lodash';
+// import * as _ from 'lodash';
 
 // import { PageHeader } from '../shared/page-header';
 // import { PageFooter } from '../shared/page-footer';
@@ -13,22 +13,21 @@ import * as _ from 'lodash';
 // import { RostersHeader } from './rosters-header';
 // import { RostersFaq } from './rosters-faq';
 // import { RostersSearch } from './rosters-search';
-import { NgIf } from '@angular/common';
-import { PageHeaderComponent } from './../shared/page-header.component';
-import { PageFooterComponent } from './../shared/page-footer.component';
-import { RostersFaqComponent } from './rosters-faq.component';
-import { RostersHeaderComponent } from './rosters-header.component';
-import { RostersMultiCampusComponent } from './rosters-multicampus.component';
-import { RostersCohortsComponent } from './rosters-cohorts.component';
-import { RostersSearchComponent } from './rosters-search.component';
+// import { NgIf } from '@angular/common';
+// import { PageHeaderComponent } from './../shared/page-header.component';
+// import { PageFooterComponent } from './../shared/page-footer.component';
+// import { RostersFaqComponent } from './rosters-faq.component';
+// import { RostersHeaderComponent } from './rosters-header.component';
+// import { RostersMultiCampusComponent } from './rosters-multicampus.component';
+// import { RostersCohortsComponent } from './rosters-cohorts.component';
+// import { RostersSearchComponent } from './rosters-search.component';
 import { AuthService } from './../../services/auth.service';
 
 
 @Component({
     selector: 'rosters',
-    providers: [],
-    templateUrl: 'components/rosters/rosters.component.html',
-    directives: [ROUTER_DIRECTIVES, PageHeaderComponent, PageFooterComponent, RostersFaqComponent, RostersHeaderComponent, RostersMultiCampusComponent, RostersCohortsComponent, RostersSearchComponent, NgIf]
+    templateUrl: './rosters.component.html',
+    // directives: [, PageHeaderComponent, PageFooterComponent, RostersFaqComponent, RostersHeaderComponent, RostersMultiCampusComponent, RostersCohortsComponent, RostersSearchComponent, NgIf]
 })
 
 export class RostersComponent implements OnInit {
