@@ -1,16 +1,12 @@
 ﻿import {Component, Input, OnInit, Output, EventEmitter, OnDestroy} from '@angular/core';
-import {ROUTER_DIRECTIVES} from '@angular/router';
 // import {CommonService} from '../../services/common';
 import {SortPipe} from '../../pipes/sort.pipe';
 import { CommonService } from './../../services/common.service';
 
 @Component({
     selector: 'self-pay-student-popup',
-    templateUrl: 'components/tests/self-pay-student-popup.component.html',
-    directives: [ROUTER_DIRECTIVES],
-    providers: [CommonService],
-    inputs: ['studentWindowException'],
-    pipes:[SortPipe]
+    templateUrl: './self-pay-student-popup.component.html',
+    inputs: ['studentWindowException']
 })
 
 export class SelfPayStudentPopupComponent implements OnInit, OnDestroy {
