@@ -1,29 +1,17 @@
-import {Component, OnInit} from '@angular/core';
-import {Router, ROUTER_DIRECTIVES, CanActivate} from '@angular/router';
-import {Title} from '@angular/platform-browser';
-// import {AuthService} from '../../services/auth';
-// import {CommonService} from '../../services/common';
-import {links} from '../../constants/config';
-// import {PageHeader} from '../shared/page-header';
-// import {TermsOfUse} from '../terms-of-use/terms-of-use';
-import {ParseDatePipe} from '../../pipes/parsedate.pipe';
-import {UtilityService} from '../../services/utility.service';
-import {Angulartics2On} from 'angulartics2';
+import { Component, OnInit } from '@angular/core';
+import { Router, CanActivate } from '@angular/router';
+import { Title } from '@angular/platform-browser';
+import { links } from '../../constants/config';
+import { UtilityService } from '../../services/utility.service';
 import { AuthService } from './../../services/auth.service';
 import { CommonService } from './../../services/common.service';
-import { PageHeaderComponent } from './../shared/page-header.component';
-import { TermsOfUseComponent } from './../terms-of-use/terms-of-use.component';
 
 @Component({
     selector: 'userguide',
-    templateUrl: 'components/userguide/userguide.component.html',
+    templateUrl: './userguide.component.html',
     host: {
         '(window:scroll)': 'onScroll($event)'
-    },
-
-    providers: [AuthService, UtilityService, CommonService],
-    directives: [PageHeaderComponent, ROUTER_DIRECTIVES, Angulartics2On, TermsOfUseComponent],
-    pipes: [ParseDatePipe]
+    }
 })
 export class UserGuideComponent implements OnInit {
     activeId: string;
@@ -34,44 +22,43 @@ export class UserGuideComponent implements OnInit {
 
     ngOnInit(): void {
         if (this.auth.isAuth()) {
-        this.titleService.setTitle('Faculty User Guide – Kaplan Nursing');
-        window.scroll(0,0);
-        this.activeId = '#whatsNew';
-        this.termsAccepted = true;
-        console.log(this.termsAccepted);
+            this.titleService.setTitle('Faculty User Guide – Kaplan Nursing');
+            window.scroll(0, 0);
+            this.activeId = '#whatsNew';
+            this.termsAccepted = true;
         }
         else {
             this.redirectToLogin();
         }
     }
-    
-     redirectToLogin() {
+
+    redirectToLogin() {
         this.router.navigate(['/']);
     }
-    
+
     //on click event added to elements in html template
     scroll(element: string, e: any) {
         e.preventDefault();
         let __this = this;
         let offset = 0;
-            offset = $('header').outerHeight(true) 
+        offset = $('header').outerHeight(true)
         if (element === '#whatsNew') {  // Includes whats New link and Back to Top link
             offset = $('header').outerHeight(true) + $('.faculty-user-guide-header').outerHeight(true);
             if ($(window).width() < 768 && $(e.target).attr('id') === 'backToTop') {
-             offset += $('.faculty-user-guide-menu ul').outerHeight(true) + 10;
+                offset += $('.faculty-user-guide-menu ul').outerHeight(true) + 10;
             }
             if ($(window).width() < 768 && $(e.target).attr('id') === 'whatsNewLink') {
-             offset = $('header').outerHeight(true) 
+                offset = $('header').outerHeight(true)
             }
-         } 
-         
+        }
+
         if ($(element).is('.h5, .h6')) { //if element is a subelement add 15px margin (may want to refactor and add specific classes)
-          offset = offset + 15;
+            offset = offset + 15;
         }
         $('html, body').animate({
             scrollTop: $(element).offset().top - offset
-        }, 500, 'swing',function () {
-             __this.activeId = element;
+        }, 500, 'swing', function () {
+            __this.activeId = element;
         });
     }
 
@@ -90,19 +77,19 @@ export class UserGuideComponent implements OnInit {
             }, 200);
         }
     }
-    
-    
+
+
     expand(element: string) {
         $(element).toggleClass('in').prev('a').toggleClass('collapsed');
 
-       if ($(element).hasClass('in')) {
-        $(element).prev('a').attr('aria-expanded', 'true');
-    } else  {
-        $(element).prev('a').attr('aria-expanded', 'false');
+        if ($(element).hasClass('in')) {
+            $(element).prev('a').attr('aria-expanded', 'true');
+        } else {
+            $(element).prev('a').attr('aria-expanded', 'false');
+        }
     }
-   }
-   
-   
+
+
     print(e: any): void {
         window.print();
         e.preventDefault();
@@ -117,4 +104,3 @@ export class UserGuideComponent implements OnInit {
     }
 
 }
-   
