@@ -1,5 +1,5 @@
-﻿import {Pipe, PipeTransform} from '@angular/core';
-import * as moment from 'moment-timezone';
+﻿import { Pipe, PipeTransform } from '@angular/core';
+
 
 @Pipe({
     name: "parseDate"
@@ -7,7 +7,7 @@ import * as moment from 'moment-timezone';
 export class ParseDatePipe implements PipeTransform {
    transform(value: any, dateFormat: string): string {
         if (!value)
-            return null;
+           return null;
         return moment(value).format(dateFormat);
     }
 }
